Move `transposeResultToOption` to `PlainOption/transpose`
We push items related to `PlainOption` into `PlainOption` modules
