+++ packages/api_tests/__tests__/plain_option/transpose/to_plain_option.test.mjs
@@ -1,8 +1,8 @@
 import test from 'ava';
 
 import { createSome, createNone, isSome, isNone } from 'option-t/PlainOption/Option';
+import { transposeResultToOption } from 'option-t/PlainOption/transpose';
 import { createOk, createErr, isOk, isErr } from 'option-t/PlainResult/Result';
-import { transposeResultToOption } from 'option-t/PlainResult/transpose';
 
 test('input is Ok<Some<T>>, the result should be Some(Ok(v))', (t) => {
     const val = Symbol('val');
