+++ packages/option-t/src/plain_result/index.ts
@@ -61,11 +61,7 @@ export { orElseAsyncForResult } from './or_else_async.js';
 export { toNullableFromErr, toNullableFromOk } from './to_nullable.js';
 // We don't export to_option since its type has been deprecated.
 export { toUndefinableFromErr, toUndefinableFromOk } from './to_undefinable.js';
-export {
-    // We don't export `transposeForResult()` from this since `PlainOption` type has been deprecated.
-    transposeResultToNullable,
-    transposeResultToUndefinable,
-} from './transpose.js';
+export { transposeResultToNullable, transposeResultToUndefinable } from './transpose.js';
 export { tryCatchIntoResult, tryCatchIntoResultWithEnsureError } from './try_catch.js';
 export {
     tryCatchIntoResultAsync,
@@ -78,11 +74,7 @@ export { unwrapOrElseAsyncFromResult } from './unwrap_or_else_async.js';
 //   that is provided only for the case to bridge with exist codes.
 //   We recommend to handle result type in that style.
 
-import {
-    // We don't export `transposeForResult()` from this since `PlainOption` type has been deprecated.
-    transposeResultToNullable,
-    transposeResultToUndefinable,
-} from './transpose.js';
+import { transposeResultToNullable, transposeResultToUndefinable } from './transpose.js';
 
 /**
  *  @deprecated
